Fold BaseFuture into PlainActionFuture (#93152)
Following #93145 and #93146, a `PlainActionFuture` is now just a
`BaseFuture` plus a few utility `actionGet(...)` methods. The only
(non-test) subclass of `BaseFuture` that doesn't inherit these methods
is `ListenableFuture`, and there's no problem with adding them there, so
this commit makes `ListenableFuture` inherit from `PlainActionFuture`
and folds the now-unnecessary `BaseFuture` into `PlainActionFuture`.
