+++ server/src/main/java/org/elasticsearch/action/support/ListenableActionFuture.java
@@ -9,14 +9,29 @@
 package org.elasticsearch.action.support;
 
 import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.common.util.concurrent.ListenableFuture;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * A {@code Future} and {@link ActionListener} against which other {@link ActionListener}s can be registered later, to support
- * fanning-out a result to a dynamic collection of listeners.
+ * An {@link ActionListener} which allows for the result to fan out to a (dynamic) collection of other listeners, added using {@link
+ * #addListener}. Listeners added before completion are retained until completion; listeners added after completion are completed
+ * immediately.
+ *
+ * Similar to {@link ListenableFuture}, with the following differences:
+ *
+ * <ul>
+ * <li>This listener will silently ignore additional completions, whereas {@link ListenableFuture} must not be completed more than once.
+ * <li>This listener completes the retained listeners on directly the completing thread, so you must use {@link ThreadedActionListener} if
+ * dispatching is needed. In contrast, {@link ListenableFuture} allows to dispatch only the retained listeners, while immediately-completed
+ * listeners are completed on the subscribing thread.
+ * <li>This listener completes the retained listeners in the context of the completing thread, so you must remember to use {@link
+ * ContextPreservingActionListener} to capture the thread context yourself if needed. In contrast, {@link ListenableFuture} allows for the
+ * thread context to be captured at subscription time.
+ * </ul>
  */
+// The name {@link ListenableActionFuture} dates back a long way and could be improved - TODO find a better name
 public class ListenableActionFuture<T> extends PlainActionFuture<T> {
 
     private Object listeners;
