+++ server/src/test/java/org/elasticsearch/threadpool/ScheduleWithFixedDelayTests.java
@@ -11,7 +11,6 @@ package org.elasticsearch.threadpool;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.support.PlainActionFuture;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.util.concurrent.BaseFuture;
 import org.elasticsearch.common.util.concurrent.EsRejectedExecutionException;
 import org.elasticsearch.core.TimeValue;
 import org.elasticsearch.node.Node;
@@ -179,8 +178,7 @@ public class ScheduleWithFixedDelayTests extends ESTestCase {
     }
 
     public void testBlockingCallOnSchedulerThreadFails() throws Exception {
-        final BaseFuture<Object> future = new BaseFuture<>() {
-        };
+        final PlainActionFuture<Object> future = new PlainActionFuture<>();
         final PlainActionFuture<Object> resultsFuture = new PlainActionFuture<>();
         final boolean getWithTimeout = randomBoolean();
 
@@ -209,13 +207,7 @@ public class ScheduleWithFixedDelayTests extends ESTestCase {
     }
 
     public void testBlockingCallOnNonSchedulerThreadAllowed() throws Exception {
-        class TestFuture extends BaseFuture<Object> {
-            public boolean set(Object o) {
-                return super.set(o);
-            }
-        }
-        final TestFuture future = new TestFuture() {
-        };
+        final PlainActionFuture<Object> future = new PlainActionFuture<>();
         final PlainActionFuture<Object> resultsFuture = new PlainActionFuture<>();
         final boolean rethrow = randomBoolean();
         final boolean getWithTimeout = randomBoolean();
@@ -241,7 +233,7 @@ public class ScheduleWithFixedDelayTests extends ESTestCase {
         assertFalse(resultsFuture.isDone());
 
         final Object o = new Object();
-        future.set(o);
+        future.onResponse(o);
 
         final Object resultingObject = resultsFuture.get();
         assertThat(resultingObject, sameInstance(o));
